+++ src/app/cause-apply/layout.tsx
@@ -1,0 +1,7 @@
+export default function DashboardLayout({
+    children, // will be a page or nested layout
+  }: {
+    children: React.ReactNode
+  }) {
+    return children
+  }
